+++ CHANGELOG.md
@@ -56,6 +56,7 @@ v2.0
 - Fix FLEXIAPI-337 Generate the provisioning URLs based on the user space
 - Fix FLEXIAPI-326 Rework email templates and translations
 - Fix FLEXIAPI-340 Fix the space resolution when getting the realm on Accounts
+- Fix FLEXIAPI-341 Allow realm to be empty when creating a Space
 
 v1.6
 ----